get rid of ValidationSet class (#736)
