+++ src/main/scala/org/tresamigos/smv/Validation.scala
@@ -95,90 +95,3 @@ private[smv] abstract class ValidationTask {
   def needAction(): Boolean
   def validate(df: DataFrame): ValidationResult
 }
-
-
-/**
- * ValidationSet is a collection of ValidationTask's
- * it provide a single entry to the list of tasks from SmvDataSet
- * TODO: should probably get rid of ValidationSet as we only do DQM validation now.
- **/
-private[smv] class ValidationSet(val tasks: Seq[ValidationTask], val isPersist: Boolean = true) {
-  def add(task: ValidationTask) = {
-    new ValidationSet(tasks :+ task, isPersist)
-  }
-
-  /** Since optimization can be done on a DF actions like count, we have to convert DF
-   *  to RDD and than apply an action
-   **/
-  private def forceAction(df: DataFrame) = {
-    df.rdd.count
-    Unit
-  }
-
-  private def terminateAtError(result: ValidationResult) = {
-    if (!result.passed) {
-      val r = result.toJSON()
-      throw new SmvDqmValidationError(r)
-    }
-  }
-
-  private def toConsole(res: ValidationResult) = {
-    SmvReportIO.printReport(res.toJSON())
-  }
-
-  private def persist(res: ValidationResult, path: String) = {
-    SmvReportIO.saveReport(res.toJSON, path)
-  }
-
-  private def readPersistsedValidationFile(path: String): Try[ValidationResult] = {
-    Try({
-      val json = SmvReportIO.readReport(path)
-      ValidationResult(json)
-    })
-  }
-
-  private def doValidate(df: DataFrame, forceAnAction: Boolean, path: String): ValidationResult = {
-    if (forceAnAction) forceAction(df)
-
-    val res = tasks
-      .map { t =>
-        t.validate(df)
-      }
-      .reduce(_ ++ _)
-
-    // as long as result non-empty, print it to the console
-    if (!res.isEmpty) toConsole(res)
-
-    // persist if result is not empty or forced an action
-    if (isPersist && ((!res.isEmpty) || forceAnAction)) persist(res, path)
-
-    res
-  }
-
-  def validate(df: DataFrame, hadAction: Boolean, path: String = ""): ValidationResult = {
-    if (!tasks.isEmpty) {
-      val needAction = tasks
-        .map { t =>
-          t.needAction
-        }
-        .reduce(_ || _)
-      val forceAnAction = (!hadAction) && needAction
-
-      val result = if (isPersist) {
-        // try to read from persisted validation file
-        readPersistsedValidationFile(path).recoverWith {
-          case e => {
-            Try(doValidate(df, forceAnAction, path))
-          }
-        }.get
-      } else {
-        doValidate(df, forceAnAction, path)
-      }
-
-      terminateAtError(result)
-      result
-    } else {
-      ValidationResult(true)
-    }
-  }
-}
