+++ src/test/scala/org/tresamigos/smv/DQMTest.scala
@@ -81,7 +81,7 @@ class DQMTest extends SmvTestUtil {
     val ssc = sqlContext; import ssc.implicits._
     val df  = dfFrom("a:Integer;b:Double", "1,0.3;0,0.2")
 
-    val dqm = new DQMValidator(SmvDQM().add(DQMRule($"a" <= 0, "a_le_0", FailAny)))
+    val dqm = new DQMValidator(SmvDQM().add(DQMRule($"a" <= 0, "a_le_0", FailAny)), false)
 
     val res = dqm.attachTasks(df)
     assert(res.count === 1)
@@ -106,7 +106,8 @@ class DQMTest extends SmvTestUtil {
     val dqm = new DQMValidator(
       SmvDQM()
         .add(DQMRule($"b" < 0.4, "b_lt_03", FailPercent(0.5)))
-        .add(DQMFix($"a" < 1, lit(1) as "a", "a_lt_1_fix", FailPercent(0.3))))
+        .add(DQMFix($"a" < 1, lit(1) as "a", "a_lt_1_fix", FailPercent(0.3))),
+      false)
 
     val res = dqm.attachTasks(df)
     assertSrddDataEqual(res, "1,0.3;1,0.2")
@@ -137,7 +138,8 @@ class DQMTest extends SmvTestUtil {
         .add(FailTotalRuleCountPolicy(2))
         .add(FailTotalFixCountPolicy(1))
         .add(FailTotalRulePercentPolicy(0.3))
-        .add(FailTotalFixPercentPolicy(0.3)))
+        .add(FailTotalFixPercentPolicy(0.3)),
+      false)
 
     val res = dqm.attachTasks(df)
